+++ application/views/ambulatorio/procedimentoplano2-lista.php
@@ -146,6 +146,9 @@ $data['empresa_permissao'] = $this->guia->listarempresapermissoes();
                             <div>
                                 <div id="circulo" style="background-color: blue"></div> Ajuste
                             </div>
+                            <div>
+                                <div id="circulo" style="background-color: gray"></div> Desativ.
+                            </div>
                         </th>
                     </tr>
                 </form>
@@ -167,15 +170,17 @@ $data['empresa_permissao'] = $this->guia->listarempresapermissoes();
                     </tr>
                 </thead>
                 <?php
+                // var_dump($limite_paginacao); die;
                 $url = $this->utilitario->build_query_params(current_url(), $_GET);
                 $limit = $limite_paginacao;
+                $lista_count = $this->procedimentoplano->listar2($_GET)->get()->result();
                 isset($_GET['per_page']) ? $pagina = $_GET['per_page'] : $pagina = 0;
                 if ($limit != "todos") {
                     $lista = $this->procedimentoplano->listar2($_GET)->limit($limit, $pagina)->get()->result();
                 } else {
                     $lista = $this->procedimentoplano->listar2($_GET)->get()->result();
                 }
-                $total = count($lista);
+                $total = count($lista_count);
                 
                 
                 if ($total > 0) {
@@ -188,6 +193,7 @@ $data['empresa_permissao'] = $this->guia->listarempresapermissoes();
                         
                         $convenioAtual = '';
                         $procedimentoAtual = '';
+                        // $ativoAtual = '';
                         
                         foreach ($lista as $item) {
                             if($convenioAtual == $item->convenio_id && $procedimentoAtual == $item->procedimento_tuss_id){
@@ -196,7 +202,9 @@ $data['empresa_permissao'] = $this->guia->listarempresapermissoes();
                             }
                             $convenioAtual = $item->convenio_id;
                             $procedimentoAtual = $item->procedimento_tuss_id;
-                            
+                            // $ativoAtual = $item->ativo;
+                            $lista_ativo = $this->procedimentoplano->listarativoprocedimento($item->procedimento_tuss_id, $item->convenio_id);
+                            // var_dump($lista_ativo);die;
                             
                             ($estilo_linha == "tabela_content01") ? $estilo_linha = "tabela_content02" : $estilo_linha = "tabela_content01";
                             ?>
@@ -214,26 +222,39 @@ $data['empresa_permissao'] = $this->guia->listarempresapermissoes();
                                     } else {
                                         $cor = 'black';
                                     }
+                                    
                                     $valor = $item->valortotal;
                                     
                                     if( $item->valor_ajuste != null){
                                         $cor = 'blue';
+                                    }
+
+                                    // Verifica se todos os procedimentos dessa linha estão desativados
+                                    // Caso sim, ele mostra em cinza
+                                    if(count($lista_ativo) == 1){
+                                        if($lista_ativo[0]->ativo == 'f'){
+                                            $cor = 'gray';
+                                        }
                                     }
                                     ?>
                                     <span style="font-weight: bolder; color: <?=$cor?>">
                                         <?= number_format((float)$valor, 2, ',', ''); ?>
                                     </span>
                                 </td>
-                                <td class="<?php echo $estilo_linha; ?>" width="80px;"> 
-                                    <a href="<?php echo base_url() ?>ambulatorio/procedimentoplano/listaprocedimentomultiempresa/<?= $item->procedimento_tuss_id ?>/<?=$item->convenio_id ?>">
-                                        Detalhes
-                                    </a>
-                                </td>
-                                <td class="<?php echo $estilo_linha; ?>" width="80px;"> 
-                                    <a target="_blank" onclick="javascript:return confirm('Deseja realmente excluir esse procedimento?');" href="<?php echo base_url() ?>ambulatorio/procedimentoplano/excluirprocedimentomultiempresa/<?= $item->procedimento_tuss_id ?>/<?=$item->convenio_id ?>">
-                                        Excluir
-                                    </a>
-                                </td>
+                                
+                                    <td class="<?php echo $estilo_linha; ?>" width="80px;"> 
+                                        <a href="<?php echo base_url() ?>ambulatorio/procedimentoplano/listaprocedimentomultiempresa/<?= $item->procedimento_tuss_id ?>/<?=$item->convenio_id ?>">
+                                            Detalhes
+                                        </a>
+                                    </td>
+                               
+                                <?if($item->associado == 'f'){?>
+                                    <td class="<?php echo $estilo_linha; ?>" width="80px;"> 
+                                        <a target="_blank" onclick="javascript:return confirm('Deseja realmente excluir esse procedimento?');" href="<?php echo base_url() ?>ambulatorio/procedimentoplano/excluirprocedimentomultiempresa/<?= $item->procedimento_tuss_id ?>/<?=$item->convenio_id ?>">
+                                            Excluir
+                                        </a>
+                                    </td>
+                                <?}?>
                             </tr>
 
                         </tbody>
